requirements.txt
scikit-learn
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
Pillow==7.0.0
PySocks==1.7.1
SQLAlchemy==1.3.13
Werkzeug==1.0.0
alembic==1.4.0
basilica==0.2.8
certifi==2019.11.28
chardet==3.0.4
idna==2.8
itsdangerous==1.1.0
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
pip==19.0.3
psycopg2
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
setuptools==40.8.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
gunicorn